+++ BankingSystem/Models/Implementations/BankServices/DepositService/DefaultDeposit.cs
@@ -11,7 +11,7 @@
         /// <param name="depositBalance">баланс</param>
         /// <param name="depositCapitalization">капитализация</param>
         /// <param name="depositRate">ставка</param>
-        public DefaultDeposit(decimal depositBalance, bool depositCapitalization, decimal depositRate) : base(depositBalance, depositCapitalization, depositRate) { }
+        public DefaultDeposit(decimal depositBalance, bool depositCapitalization, double depositRate) : base(depositBalance, depositCapitalization, depositRate) { }
 
         /// <summary>
         /// Метод возвращаюший состояние объекта
